backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8

# Image handling
Pillow==10.1.0

# CORS
django-cors-headers==4.3.1

# AI & RAG
groq==0.4.1
chromadb==0.4.22
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.24.3

# Utilities
Faker==20.1.0
python-dotenv==1.0.0